Mostly working but need to abandon pandoc bindings
I stopped looking into the crowbook crate because it's mostly for books.
The pandoc crate is a little awkward, and only allows output to stdout
or a file; as it turns out, I need to use pandoc only on the context
portion of each post (rather than templating and then converting),
otherwise pandoc will try to convert valid html from markdown to html.

I kinda don't want to write posts in html, so I'm probably just going to
use a std::proccess::Command or something instead to invoke pandoc and
get handle to stdout (without just printing it to the screen).